+++ 05_BACKEND/complete backend/src/controllers/video.controller.js
@@ -1,20 +1,26 @@
 import mongoose, { isValidObjectId } from "mongoose";
 import { Video } from "../models/video.model.js";
-import { User } from "../models/user.model.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { uploadOnCloudinary } from "../utils/cloudinary.js";
-import { title } from "process";
 
 const getAllVideos = asyncHandler(async (req, res) => {
     //TODO: get all videos based on query, sort, pagination
     const { page = 1, limit = 10, query, sortBy, sortType, userId } = req.query;
 
-    const filters = {
-        isPublished: true,
-        title: { $regex: query, $options: "i" },
-    };
+    let filters = {}
+    if (query) {
+        const filters = {
+            $or: [
+                {
+                    isPublished: true,
+                    title: { $regex: query, $options: "i" },
+                    description: { $regex: query, $options: "i" },
+                },
+            ],
+        };
+    }
 
     if (userId && isValidObjectId(userId)) {
         filters.owner = userId;
@@ -27,8 +33,15 @@ const getAllVideos = asyncHandler(async (req, res) => {
     const videos = await Video.find(filters)
         .populate("owner")
         .sort(sortOptions)
-        .skip((page - 1) * limit)
+        .skip(parseInt((page - 1) * limit))
         .limit(parseInt(limit));
+
+    const totalVideos = await Video.countDocuments(filters);
+
+    if (totalVideos.length === 0) {
+        throw new ApiError(404, "Videos not found");
+    }
+
     return res
         .status(200)
         .json(ApiResponse(200, "Videos fetched successfully", videos));
@@ -37,26 +50,34 @@ const getAllVideos = asyncHandler(async (req, res) => {
 const publishAVideo = asyncHandler(async (req, res) => {
     // TODO: get video, upload to cloudinary, create video
     const { title, description } = req.body;
-    if (!req.files?.video) {
+    if (!title || !description || !req.files?.video) {
         throw new ApiError(400, "Video is not found");
     }
 
-    const videoUpload = await uploadOnCloudinary(req.files.video.tempFilePath);
-    const thumbnailUpload = req.files.thumbnail
-        ? await uploadOnCloudinary(req.files.thumbnail.tempFilePath)
-        : null;
+    const videoFile = req.files.video;
+    const thumbnailFile = req.files.thumbnail;
 
-    const video = await Video.create({
-        title,
-        description,
-        video: videoUpload.url,
-        thumbnail: thumbnailUpload?.url || "",
-        owner: req.user._id,
-    });
+    try {
+        const videoUpload = await uploadOnCloudinary(videoFile.tempFilePath);
+        const thumbnailUpload = thumbnailFile.tempFilePath
+            ? await uploadOnCloudinary(thumbnailFile.tempFilePath)
+            : null;
 
-    return res
-        .status(200)
-        .json(ApiResponse(200, "Video published successfully", video));
+        const video = await Video.create({
+            title,
+            description,
+            video: videoUpload.url,
+            thumbnail: thumbnailUpload?.url || "",
+            owner: req.user._id,
+            isPublished: false,
+        });
+
+        return res
+            .status(200)
+            .json(ApiResponse(200, "Video published successfully", video));
+    } catch (error) {
+        throw new ApiError(500, "Failed to publish video");
+    }
 });
 
 const getVideoById = asyncHandler(async (req, res) => {
@@ -88,6 +109,11 @@ const updateVideo = asyncHandler(async (req, res) => {
     if (!video) {
         throw new ApiError(404, "Video not found");
     }
+
+    if (req.user._id.toString() !== video.owner.toString()) {
+        throw new ApiError(403, "You are not authorized to update this video");
+    }
+
     if (title) {
         video.title = title;
     }
@@ -114,7 +140,12 @@ const deleteVideo = asyncHandler(async (req, res) => {
         throw new ApiError(400, "Invalid video id");
     }
 
+    if (req.user._id.toString() !== video.owner.toString()) {
+        throw new ApiError(403, "You are not authorized to delete this video");
+    }
+
     const video = await Video.findByIdAndDelete(videoId);
+    await uploadOnCloudinary.deleteResource(video.cloudinaryPublicId);
     if (!video) {
         throw new ApiError(404, "Video not found");
     }
@@ -134,6 +165,10 @@ const togglePublishStatus = asyncHandler(async (req, res) => {
     const video = await Video.findById(videoId);
     if (!video) {
         throw new ApiError(404, "Video not found");
+    }
+
+    if (req.user._id.toString() !== video.owner.toString()) {
+        throw new ApiError(403, "You are not authorized to update this video");
     }
 
     video.isPublished = !video.isPublished;
